pyproject.toml
[project]
name = "ag-ui-deepchat-poc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.4.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
